refactor(wip): typemanagers also fix bad operator!=
