+++ include/private/logics/typemanager.h
@@ -5,6 +5,8 @@
 #include <QJsonObject>
 #include <QMap>
 
+#include "qtgraph.h"
+
 #include "typemanager.pb.h"
 
 namespace qtgraph {
@@ -22,7 +24,11 @@ public:
 
     inline QString typeNameByID(int id) const { return Types()[id].value("name").toString(); }
 
-    virtual bool readTypes(const char *file) = 0;
+    bool readTypes(const char *file);
+    inline bool readTypes(const QString &file) { return readTypes(file.toStdString().c_str()); }
+    inline bool readTypes(const std::string &file) { return readTypes(file.c_str()); }
+
+    virtual QString keyValue() const = 0;
 
     void setFileName(QString name) { _filename = name; }
     const QString &getFileName() const { return _filename; }
@@ -37,6 +43,26 @@ protected:
     QMap<QString, int> _typeNames = {};
     QString _filename = "";
 
+};
+
+class PinTypeManager final : public TypeManager
+{
+public:
+    PinTypeManager() {}
+
+    QString keyValue() const override { return "pins"; }
+
+    static PinTypeManager *fromProtocolTypeManager(const protocol::TypeManager &tm);
+};
+
+class NodeTypeManager final : public TypeManager
+{
+public:
+    NodeTypeManager() {}
+
+    QString keyValue() const override { return "nodes"; }
+
+    static NodeTypeManager *fromProtocolTypeManager(const protocol::TypeManager &tm);
 };
 
 }
